backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.9
alembic>=1.14.0
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
redis>=5.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
httpx>=0.27.0
pytest>=8.3.0
